refactor: exception 위치 수정 및 handler 작성
