+++ src/main/java/org/cookieandkakao/babting/domain/calendar/exception/EventCreationException.java
@@ -1,4 +1,4 @@
-package org.cookieandkakao.babting.common.exception.customexception;
+package org.cookieandkakao.babting.domain.calendar.exception;
 
 public class EventCreationException extends RuntimeException {
     public EventCreationException(String message) {
