+++ src/main/java/org/cookieandkakao/babting/domain/calendar/exception/JsonConversionException.java
@@ -1,4 +1,4 @@
-package org.cookieandkakao.babting.common.exception.customexception;
+package org.cookieandkakao.babting.domain.calendar.exception;
 
 public class JsonConversionException extends RuntimeException {
     public JsonConversionException(String message) {
