+++ src/main/java/org/cookieandkakao/babting/domain/calendar/service/TalkCalendarService.java
@@ -9,8 +9,8 @@ import java.util.List;
 import java.util.Optional;
 import org.cookieandkakao.babting.common.cache.CacheKeyGenerator;
 import org.cookieandkakao.babting.common.exception.customexception.CacheEvictionException;
-import org.cookieandkakao.babting.common.exception.customexception.EventCreationException;
-import org.cookieandkakao.babting.common.exception.customexception.JsonConversionException;
+import org.cookieandkakao.babting.domain.calendar.exception.EventCreationException;
+import org.cookieandkakao.babting.domain.calendar.exception.JsonConversionException;
 import org.cookieandkakao.babting.domain.calendar.dto.request.EventCreateRequest;
 import org.cookieandkakao.babting.domain.calendar.dto.response.EventCreateResponse;
 import org.cookieandkakao.babting.domain.calendar.dto.response.EventDetailGetResponse;
