+++ src/main/java/org/cookieandkakao/babting/domain/food/exception/FoodNotFoundException.java
@@ -1,4 +1,4 @@
-package org.cookieandkakao.babting.common.exception.customexception;
+package org.cookieandkakao.babting.domain.food.exception;
 
 public class FoodNotFoundException extends RuntimeException {
   public FoodNotFoundException(String message) {
