+++ src/main/java/org/cookieandkakao/babting/domain/food/exception/InvalidFoodPreferenceTypeException.java
@@ -1,4 +1,4 @@
-package org.cookieandkakao.babting.common.exception.customexception;
+package org.cookieandkakao.babting.domain.food.exception;
 
 public class InvalidFoodPreferenceTypeException extends RuntimeException {
   public InvalidFoodPreferenceTypeException(String message) {
