+++ src/main/java/org/cookieandkakao/babting/domain/food/exception/PreferenceConflictException.java
@@ -1,4 +1,4 @@
-package org.cookieandkakao.babting.common.exception.customexception;
+package org.cookieandkakao.babting.domain.food.exception;
 
 public class PreferenceConflictException extends RuntimeException {
   public PreferenceConflictException(String message) {
