+++ src/main/java/org/cookieandkakao/babting/domain/member/exception/MemberNotFoundException.java
@@ -1,4 +1,4 @@
-package org.cookieandkakao.babting.common.exception.customexception;
+package org.cookieandkakao.babting.domain.member.exception;
 
 public class MemberNotFoundException extends RuntimeException {
     public MemberNotFoundException(String message) {
